Replace 3.8.5 Python version with 3.8.10 in data for consistency
I had to update Python to 3.8.10 to install some packages, and this
screwed up baselines which were based on 3.8.5. Since only the point
release was changed, keep the old 3.8.5 baselines (and other results)
but adjust the version to 3.8.10 so that there will no version
conflicts. The results might be very slightly off, but it's probably
okay.